requirements.txt
flask==1.1.2
pyyaml==5.3.1
mongoengine==0.20.0
hdfs==2.5.8